requirements.txt
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
PyQt6==6.6.1
faker==20.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
bcrypt==4.1.1
